+++ src/boomerang/db/statements/JunctionStatement.cpp
@@ -65,7 +65,7 @@ void JunctionStatement::print(QTextStream& os, bool html) const
 
     os << "JUNCTION ";
 
-    for (size_t i = 0; i < m_parent->getNumPredecessors(); i++) {
+    for (int i = 0; i < m_parent->getNumPredecessors(); i++) {
         os << m_parent->getPredecessors()[i]->getHiAddr();
 
         if (m_parent->isBackEdge(i)) {
@@ -89,7 +89,7 @@ void JunctionStatement::print(QTextStream& os, bool html) const
 
 bool JunctionStatement::isLoopJunction() const
 {
-    for (size_t i = 0; i < m_parent->getNumPredecessors(); i++) {
+    for (int i = 0; i < m_parent->getNumPredecessors(); i++) {
         if (m_parent->isBackEdge(i)) {
             return true;
         }
